fix: rename models to store and resolve migration glob issue
