+++ blocklets/image-bin/api/hooks/pre-start.js
@@ -1,12 +1,11 @@
-const { migrate } = require('../migrations');
-const logger = require('../libs/logger');
+const { migrate } = require('../store/migrate');
 
 (async () => {
   try {
     await migrate();
     process.exit(0);
   } catch (err) {
-    logger.error('[pre-start] failed', err);
+    console.error('[pre-start] failed', err);
     process.exit(1);
   }
 })();
