+++ blocklets/image-bin/api/store/migrate.js
@@ -1,10 +1,14 @@
+/* eslint-disable global-require */
 const { Umzug, SequelizeStorage } = require('umzug');
-const { sequelize } = require('../models');
+const { sequelize } = require('./index');
 
 const umzug = new Umzug({
   migrations: {
-    glob: ['*.js', { cwd: __dirname, ignore: ['**/index.js'] }],
+    glob: ['**/migrations/*.{ts,js}', { cwd: __dirname }],
+    // migrations: [require('./migrations/2025060401-genesis'), require('./migrations/2025060402-migrate-from-nedb')],
     resolve: ({ name, path, context }) => {
+      // eslint-disable-next-line no-console
+      console.log('migrations path', path);
       // eslint-disable-next-line import/no-dynamic-require, global-require
       const migration = require(path);
       return {
@@ -19,10 +23,8 @@ const umzug = new Umzug({
   logger: console,
 });
 
-// Export the initialized umzug instance
 module.exports.umzug = umzug;
 
-// Export a function to run migrations
 module.exports.migrate = async () => {
   await umzug.up();
 };
