+++ blocklets/image-bin/package.json
@@ -10,7 +10,7 @@
     "clean": "rm -rf .blocklet",
     "bundle": "npm run bundle:client && npm run bundle:api",
     "bundle:client": "npm run build",
-    "bundle:api": "npm run clean && blocklet bundle --compact --create-release --source-map --no-minify",
+    "bundle:api": "pnpm run clean && blocklet bundle --compact --monorepo --create-release --changelog",
     "deploy": "npm run bundle && blocklet deploy .blocklet/bundle",
     "upload": "npm run bundle && blocklet upload .blocklet/release/blocklet.json",
     "update:deps": "npx taze -n '/arcblock|ocap|abtnode|blocklet|did-connect|did-comment|nedb/' -w -f && pnpm i && pnpm run deduplicate",
